+++ src/firecracker/src/api_server/request/balloon.rs
@@ -1,7 +1,7 @@
 // Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-use micro_http::StatusCode;
+use micro_http::{Method, StatusCode};
 use vmm::rpc_interface::VmmAction;
 use vmm::vmm_config::balloon::{
     BalloonDeviceConfig, BalloonUpdateConfig, BalloonUpdateStatsConfig,
@@ -9,18 +9,36 @@ use vmm::vmm_config::balloon::{
 
 use super::super::parsed_request::{ParsedRequest, RequestError};
 use super::Body;
+use crate::api_server::parsed_request::method_to_error;
 
-pub(crate) fn parse_get_balloon(
-    path_second_token: Option<&str>,
-) -> Result<ParsedRequest, RequestError> {
-    match path_second_token {
-        Some(stats_path) => match stats_path {
-            "statistics" => Ok(ParsedRequest::new_sync(VmmAction::GetBalloonStats)),
-            _ => Err(RequestError::Generic(
-                StatusCode::BadRequest,
-                format!("Unrecognized GET request path `{}`.", stats_path),
-            )),
-        },
+fn parse_get_hinting<'a, T>(mut path_tokens: T) -> Result<ParsedRequest, RequestError>
+where
+    T: Iterator<Item = &'a str>,
+{
+    match path_tokens.next() {
+        Some("status") => Ok(ParsedRequest::new_sync(VmmAction::GetFreePageHintingStatus)),
+        Some(stats_path) => Err(RequestError::Generic(
+            StatusCode::BadRequest,
+            format!("Unrecognized GET request path `/hinting/{stats_path}`."),
+        )),
+        None => Err(RequestError::Generic(
+            StatusCode::BadRequest,
+            "Unrecognized GET request path `/hinting/`.".to_string(),
+        )),
+    }
+}
+
+pub(crate) fn parse_get_balloon<'a, T>(mut path_tokens: T) -> Result<ParsedRequest, RequestError>
+where
+    T: Iterator<Item = &'a str>,
+{
+    match path_tokens.next() {
+        Some("statistics") => Ok(ParsedRequest::new_sync(VmmAction::GetBalloonStats)),
+        Some("hinting") => parse_get_hinting(path_tokens),
+        Some(stats_path) => Err(RequestError::Generic(
+            StatusCode::BadRequest,
+            format!("Unrecognized GET request path `{}`.", stats_path),
+        )),
         None => Ok(ParsedRequest::new_sync(VmmAction::GetBalloonConfig)),
     }
 }
@@ -31,23 +49,55 @@ pub(crate) fn parse_put_balloon(body: &Body) -> Result<ParsedRequest, RequestErr
     )))
 }
 
-pub(crate) fn parse_patch_balloon(
-    body: &Body,
-    path_second_token: Option<&str>,
-) -> Result<ParsedRequest, RequestError> {
-    match path_second_token {
-        Some(config_path) => match config_path {
-            "statistics" => Ok(ParsedRequest::new_sync(VmmAction::UpdateBalloonStatistics(
+fn parse_patch_hinting<'a, T>(
+    body: Option<&Body>,
+    mut path_tokens: T,
+) -> Result<ParsedRequest, RequestError>
+where
+    T: Iterator<Item = &'a str>,
+{
+    match path_tokens.next() {
+        Some("start") => {
+            let cmd = match body {
+                None => Default::default(),
+                Some(b) if b.is_empty() => Default::default(),
+                Some(b) => serde_json::from_slice(b.raw())?,
+            };
+
+            Ok(ParsedRequest::new_sync(VmmAction::StartFreePageHinting(
+                cmd,
+            )))
+        }
+        Some("stop") => Ok(ParsedRequest::new_sync(VmmAction::StopFreePageHinting)),
+        Some(stats_path) => Err(RequestError::Generic(
+            StatusCode::BadRequest,
+            format!("Unrecognized PATCH request path `/hinting/{stats_path}`."),
+        )),
+        None => Err(RequestError::Generic(
+            StatusCode::BadRequest,
+            "Unrecognized PATCH request path `/hinting/`.".to_string(),
+        )),
+    }
+}
+
+pub(crate) fn parse_patch_balloon<'a, T>(
+    body: Option<&Body>,
+    mut path_tokens: T,
+) -> Result<ParsedRequest, RequestError>
+where
+    T: Iterator<Item = &'a str>,
+{
+    match (path_tokens.next(), body) {
+        (Some("statistics"), Some(body)) => {
+            Ok(ParsedRequest::new_sync(VmmAction::UpdateBalloonStatistics(
                 serde_json::from_slice::<BalloonUpdateStatsConfig>(body.raw())?,
-            ))),
-            _ => Err(RequestError::Generic(
-                StatusCode::BadRequest,
-                format!("Unrecognized PATCH request path `{}`.", config_path),
-            )),
-        },
-        None => Ok(ParsedRequest::new_sync(VmmAction::UpdateBalloon(
+            )))
+        }
+        (Some("hinting"), body) => parse_patch_hinting(body, path_tokens),
+        (_, Some(body)) => Ok(ParsedRequest::new_sync(VmmAction::UpdateBalloon(
             serde_json::from_slice::<BalloonUpdateConfig>(body.raw())?,
         ))),
+        (_, None) => method_to_error(Method::Patch),
     }
 }
 
@@ -58,30 +108,34 @@ mod tests {
 
     #[test]
     fn test_parse_get_balloon_request() {
-        parse_get_balloon(None).unwrap();
+        parse_get_balloon([].into_iter()).unwrap();
 
-        parse_get_balloon(Some("unrelated")).unwrap_err();
+        parse_get_balloon(["unrelated"].into_iter()).unwrap_err();
 
-        parse_get_balloon(Some("statistics")).unwrap();
+        parse_get_balloon(["statistics"].into_iter()).unwrap();
+
+        parse_get_balloon(["hinting", "status"].into_iter()).unwrap();
+        parse_get_balloon(["hinting", "unrelated"].into_iter()).unwrap_err();
+        parse_get_balloon(["hinting"].into_iter()).unwrap_err();
     }
 
     #[test]
     fn test_parse_patch_balloon_request() {
-        parse_patch_balloon(&Body::new("invalid_payload"), None).unwrap_err();
+        parse_patch_balloon(Some(&Body::new("invalid_payload")), [].into_iter()).unwrap_err();
 
         // PATCH with invalid fields.
         let body = r#"{
             "amount_mib": "bar",
             "foo": "bar"
         }"#;
-        parse_patch_balloon(&Body::new(body), None).unwrap_err();
+        parse_patch_balloon(Some(&Body::new(body)), [].into_iter()).unwrap_err();
 
         // PATCH with invalid types on fields. Adding a polling interval as string instead of bool.
         let body = r#"{
             "amount_mib": 1000,
             "stats_polling_interval_s": "false"
         }"#;
-        let res = parse_patch_balloon(&Body::new(body), None);
+        let res = parse_patch_balloon(Some(&Body::new(body)), [].into_iter());
         res.unwrap_err();
 
         // PATCH with invalid types on fields. Adding a amount_mib as a negative number.
@@ -89,21 +143,21 @@ mod tests {
             "amount_mib": -1000,
             "stats_polling_interval_s": true
         }"#;
-        let res = parse_patch_balloon(&Body::new(body), None);
+        let res = parse_patch_balloon(Some(&Body::new(body)), [].into_iter());
         res.unwrap_err();
 
         // PATCH on statistics with missing ppolling interval field.
         let body = r#"{
             "amount_mib": 100
         }"#;
-        let res = parse_patch_balloon(&Body::new(body), Some("statistics"));
+        let res = parse_patch_balloon(Some(&Body::new(body)), ["statistics"].into_iter());
         res.unwrap_err();
 
         // PATCH with missing amount_mib field.
         let body = r#"{
             "stats_polling_interval_s": 0
         }"#;
-        let res = parse_patch_balloon(&Body::new(body), None);
+        let res = parse_patch_balloon(Some(&Body::new(body)), [].into_iter());
         res.unwrap_err();
 
         // PATCH that tries to update something else other than allowed fields.
@@ -111,27 +165,29 @@ mod tests {
             "amount_mib": "dummy_id",
             "stats_polling_interval_s": "dummy_host"
         }"#;
-        let res = parse_patch_balloon(&Body::new(body), None);
+        let res = parse_patch_balloon(Some(&Body::new(body)), [].into_iter());
         res.unwrap_err();
 
         // PATCH with payload that is not a json.
         let body = r#"{
             "fields": "dummy_field"
         }"#;
-        parse_patch_balloon(&Body::new(body), None).unwrap_err();
+        parse_patch_balloon(Some(&Body::new(body)), [].into_iter()).unwrap_err();
 
         // PATCH on unrecognized path.
         let body = r#"{
             "fields": "dummy_field"
         }"#;
-        parse_patch_balloon(&Body::new(body), Some("config")).unwrap_err();
+        parse_patch_balloon(Some(&Body::new(body)), ["config"].into_iter()).unwrap_err();
 
         let body = r#"{
             "amount_mib": 1
         }"#;
         let expected_config = BalloonUpdateConfig { amount_mib: 1 };
         assert_eq!(
-            vmm_action_from_request(parse_patch_balloon(&Body::new(body), None).unwrap()),
+            vmm_action_from_request(
+                parse_patch_balloon(Some(&Body::new(body)), [].into_iter()).unwrap()
+            ),
             VmmAction::UpdateBalloon(expected_config)
         );
 
@@ -143,10 +199,44 @@ mod tests {
         };
         assert_eq!(
             vmm_action_from_request(
-                parse_patch_balloon(&Body::new(body), Some("statistics")).unwrap()
+                parse_patch_balloon(Some(&Body::new(body)), ["statistics"].into_iter()).unwrap()
             ),
             VmmAction::UpdateBalloonStatistics(expected_config)
         );
+
+        // PATCH start hinting run valid data
+        let body = r#"{
+            "acknowledge_on_stop": true
+        }"#;
+        parse_patch_balloon(Some(&Body::new(body)), ["hinting", "start"].into_iter()).unwrap();
+
+        // PATCH start hinting run no body
+        parse_patch_balloon(Some(&Body::new("")), ["hinting", "start"].into_iter()).unwrap();
+
+        // PATCH start hinting run invalid data
+        let body = r#"{
+            "acknowledge_on_stop": "not valid"
+        }"#;
+        parse_patch_balloon(Some(&Body::new(body)), ["hinting", "start"].into_iter()).unwrap_err();
+
+        // PATCH start hinting run no body
+        parse_patch_balloon(Some(&Body::new(body)), ["hinting", "start"].into_iter()).unwrap_err();
+
+        // PATCH stop hinting run
+        parse_patch_balloon(Some(&Body::new("")), ["hinting", "stop"].into_iter()).unwrap();
+
+        // PATCH stop hinting run
+        parse_patch_balloon(None, ["hinting", "stop"].into_iter()).unwrap();
+
+        // PATCH stop hinting invalid path
+        parse_patch_balloon(Some(&Body::new("")), ["hinting"].into_iter()).unwrap_err();
+
+        // PATCH stop hinting invalid path
+        parse_patch_balloon(Some(&Body::new("")), ["hinting", "other path"].into_iter())
+            .unwrap_err();
+
+        // PATCH no body non hinting
+        parse_patch_balloon(None, ["hinting"].into_iter()).unwrap_err();
     }
 
     #[test]
@@ -160,11 +250,29 @@ mod tests {
         }"#;
         parse_put_balloon(&Body::new(body)).unwrap_err();
 
-        // PUT with valid input fields.
+        // PUT with valid input fields. Hinting reporting missing
         let body = r#"{
             "amount_mib": 1000,
             "deflate_on_oom": true,
             "stats_polling_interval_s": 0
+        }"#;
+        parse_put_balloon(&Body::new(body)).unwrap();
+
+        // PUT with valid input hinting
+        let body = r#"{
+            "amount_mib": 1000,
+            "deflate_on_oom": true,
+            "stats_polling_interval_s": 0,
+            "free_page_hinting": true
+        }"#;
+        parse_put_balloon(&Body::new(body)).unwrap();
+
+        // PUT with valid reporting
+        let body = r#"{
+            "amount_mib": 1000,
+            "deflate_on_oom": true,
+            "stats_polling_interval_s": 0,
+            "free_page_reporting": true
         }"#;
         parse_put_balloon(&Body::new(body)).unwrap();
     }