+++ src/vmm/src/builder.rs
@@ -1305,6 +1305,8 @@ pub(crate) mod tests {
             amount_mib: 0,
             deflate_on_oom: false,
             stats_polling_interval_s: 0,
+            free_page_hinting: false,
+            free_page_reporting: false,
         };
 
         let mut cmdline = default_kernel_cmdline();